move sgir tests to a separate file
